requirements.txt
Django==5.2.4
Pillow==10.4.0
django-cors-headers==4.3.1
djangorestframework==3.15.0 